Switched to buffer and localleader
I switched to using nnoremap,
and only defining the commands for the current buffer.

Since ftplugins are only run once you know that the filetype
is tex, go ahead an define the mappings unconditionally,
but use <buffer> so that they only get loaded for the current (tex)
buffer, not all open buffers.

Also, this is a filetype specific command, so it makes sense
to use <localleader> (which is mapped to ,) anyways,
so now people who remap local leader are happy too.

Also, I split the commands up by line, that should
make it easier to see what's going on.  It also
highlights some of the redundancy, perhaps there should
be a function defined?